+++ portal/portal/templatetags/accordion_menu_helpers.py
@@ -1,7 +1,4 @@
 from django import template
-from django.utils.safestring import mark_safe
-
-from wagtail.core.models import Page
 
 
 register = template.Library()
@@ -9,7 +6,7 @@ register = template.Library()
 
 class AccordionMenuEntry:
     def __init__(self, a_wagtail_page, current_page):
-        self._page_title = str(a_wagtail_page) or ""
+        self._page_title = str(a_wagtail_page.homepage.translated_title) or ""
         self._page_url = a_wagtail_page.url if a_wagtail_page is not None else ""
         self._children = []
         self._is_active = False
@@ -50,21 +47,15 @@ class AccordionMenuEntry:
         if a_wagtail_page.numchild > 0:
             for child in a_wagtail_page.get_children().live().in_menu():
                 accordion_child = AccordionMenuEntry(child, current_page)
-                self._is_active = accordion_child.active
+                self._is_active = self.active or accordion_child.active
                 self.add_child(accordion_child)
 
 
-
-@register.filter
-def is_page_active(accordion_entry, wagtail_page):
-    return accordion_entry.title == str(wagtail_page)
-
-
-@register.inclusion_tag('home/accordion_menu.html', takes_context=True)
-def create_accordion_menu_from_site_map(context, current_page):
+@register.simple_tag(takes_context=True)
+def get_accordion_menu(context, current_page):
     wagtail_site_root = context['request'].site.root_page
     accordion_menu = AccordionMenuEntry(wagtail_site_root, current_page)
     return {
         'current_page': current_page,
-        'pages': accordion_menu,
+        'tree': accordion_menu,
     }