+++ portal/webpack.config.js
@@ -4,7 +4,9 @@ const path = require('path');
 module.exports = {
     'entry': path.resolve(__dirname, 'src', 'index.js'),
     'output': {
-        'filename': 'main.js'
+        'filename': 'main.js',
+        'library': 'myeqip',
+        'libraryTarget': 'window'
     },
     'module': {
         'rules': [