+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 ### Interface
 
 - [Footer content](https://github.com/catalpainternational/myeqip-portal/pull/31)
+- [RichTextBlock](https://github.com/catalpainternational/myeqip-portal/pull/35)
+- [Splits menu into two `nav`s.](https://github.com/catalpainternational/myeqip-portal/pull/38)
+- [QuoteBlock](https://github.com/catalpainternational/myeqip-portal/pull/40)
+- [PostcardBlock](https://github.com/catalpainternational/myeqip-portal/pull/42)
 
 ### QA